relaese version 5

+++ services/equiment/sql/delete-equiment.php
@@ -1,15 +1,20 @@
 <?php
 include '../../../connection.php';
-    $id = $_GET['id'];
+@$data = json_decode(file_get_contents("php://input"));
+@$x=count($data);
+if($x > 0) {
+    $id    = $data->id;
     $selectMenu=$_SQL($con,"SELECT e_img FROM spa_equiment WHERE e_id='$id'");
     $res=$_ASSOC($selectMenu);
     $img=$res['e_img'];
     $dir='../../../img/';
     $query = "DELETE FROM spa_equiment WHERE e_id='$id'";
     if (mysqli_query($con, $query)) {
-        echo 7070;
-        @unlink($dir.$img);
+         @unlink($dir.$img);
+        echo 200;
     } else {
-        echo 4466;
+        echo 400;
     }
+}
+mysqli_close($con);
 ?>
